Extract tests for Python 3.9 compatibility into their own file
